check-run-action: require owner/repository inputs
In case the installation token is no longer valid, we need to generate
a new one. And if we do not have an owner and repository to work with,
we will be greeted by a message like this:

  FAILED GitHub REST API call!
  {
    port: 443,
    hostname: 'api.github.com',
    method: 'GET',
    path: '/repos/git-for-windows/undefined/installation',
    headers: {
      'User-Agent': 'GitForWindowsHelperApp/0.0',
      Accept: 'application/json',
      Authorization: '***'
    }
  }

See e.g. https://github.com/git-for-windows/git-for-windows-automation/actions/runs/5001089177/jobs/8959839096#step:16:92

Let's avoid this by always requiring the owner and repository input
parameters to be provided.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>