git-artifacts: try creating `MINGW-packages.bundle` a couple times if needed
In private tests, this command seems to have failed consistently, with
this error message:

    + git commit -s -m 'mingw-w64-git: new version (test)' PKGBUILD
    [test e641948] mingw-w64-git: new version (test)
    1 file changed, 2 insertions(+), 2 deletions(-)
    + git bundle create /d/a/rss-test/rss-test/artifacts/MINGW-packages.bundle 'test^..test'
    error: pack-objects died
    Error: Process completed with exit code 1.

After extensive debugging of this elusive problem (it is relatively
consistent when running inside a workflow step, but running it
interactively via `action-tmate` succeeds every single time), the most
plausible explanation is that Defender doing its thing, and that results
in a "Permission Denied" problem.

Calling the command with `GIT_TRACE2_EVENT` enabled seems to work around
the problem, and just in case it does not, let's add a loop to try a
couple more times should the command fail.

In the end, just validate the bundle to make sure that it was created
correctly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>