git-artifacts: report selected artifacts in the workflow run's name
The name given to a particular workflow run is shown in the overview of
workflow runs in GitHub's UI, and are therefore a really good spot to
surface information that is relevant to the reader.

For example, if a `git-artifacts` workflow run was started with only a
subset of the artifacts to build, let's state that in the workflow run's
name, as it is vital information.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>