Be more careful `updateRepositoryUpdate()` handling the `build-extra` updates
The `build-extra` updates commit some changes before pushing. Since we
want to use this code also for embargoed releases, we need to decouple
this so that a "phase 1" workflow run can commit these changes, and a
"phase 2" workflow run can push the changes to the non-embargoed
repository.

Technically, this is still not quite correct: The last part, the one
updating `download-stats.sh`, can only be done once the embargo has been
lifted (because it required the ID of the public GitHub Release, which
will only be available at that time).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>