TO-DROP: Duplicate release-git
It would be better if we could somehow refactor release-git to handle
also embargoed releases.

One option would be to turn the i686/x86_64 run IDs into a JSON
structure that is passed around, and handled by a class that at first
does precisely what is being done right now, just with a much better
interface: a getter that retrieves (from the appropriate source) an
embargoed build artifact of a specific name (such as
`installer-x86_64`). Then add a commit that extends this to using a
GitHub Release in the embargoed `git` repository as source, instead of
a pair of `git-for-windows-automation` workflow runs' build artifacts.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>